Add UserPanel View And Create Component for show Infos

+++ Shop.Persistence/Repositories/Generics/GenericRepository.cs
@@ -28,7 +28,7 @@ namespace Shop.Persistence.Repositories.Generics
            return await _context.Set<T>().ToListAsync();
         }
 
-        public async Task<T> GetAsync(int id)
+        public async Task<T> GetAsync(long id)
         {
             return await _context.Set<T>().FindAsync(id);
         }
